client/requirements.txt
PyQt6>=6.4.0
opencv-python>=4.7.0
numpy>=1.24.0
websockets>=11.0.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-qt>=4.2.0 